// actions/filters/types.ts
export enum FiltersActionType {
  TOGGLE_EXCLUDED_PRESENCE = "TOGGLE_EXCLUDED",
  TOGGLE_ID_PRESENCE = "TOGGLE_ID_PRESENCE",
  TOGGLE_MULTIPLE_ID_PRESENCE = "TOGGLE_MULTIPLE_ID_PRESENCE",
  TOGGLE_STATUS_PRESENCE = "TOGGLE_STATUS",
  SET_SEARCH_TERM = "SET_SEARCH_TERM",
  TOGGLE_SEARCHABLE_FILTER_PRESENCE = "TOGGLE_SEARCHABLE_FILTER_PRESENCE",
  TOGGLE_MULTIPLE_SEARCHABLE_FILTER_PRESENCE = "TOGGLE_MULTIPLE_SEARCHABLE_FILTER_PRESENCE",
  RESET_ALL_FILTERS = "RESET_ALL_FILTERS",
  SET_USER_DATA_FILTER = "SET_USER_DATA_FILTER",
}